% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_weather.R
\name{current_weather}
\alias{current_weather}
\title{Read the Data of One Airport Location}
\usage{
current_weather(id,type)
}
\arguments{
\item{id}{one 4 digits code of the airport}

\item{type}{a vector of weather elements}
}
\value{
a data frame of the weather information
}
\description{
This function read the the weather information of a airport from the website of NOAA's National Weather Service.
}
\examples{
current_weather("KAMW", c("wind_mph", "temp_f", "haha"))
current_weather("KAMW", "temp_c")
}
