% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position.R
\name{plot_position}
\alias{plot_position}
\title{Plot an Element of Selected Airports and Your location}
\usage{
plot_position(id_vector, type, you.long, you.lat, number = T)
}
\arguments{
\item{id_vector}{a vector of  multiple airport codes}

\item{type}{a vector of weather elements}

\item{you.long}{The longitude of your location}

\item{you.lat}{The latitude of your location}
}
\description{
This function plot an element of selected airport and your position.
}
\details{
number = true: this element is numeric
}
\examples{
plot_position(c("KAMW", "KAIO"),
type = "temp_f", you.long = -100, you.lat = 35,number = TRUE)
}
