% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather_cat.R
\name{plot_weather_cat}
\alias{plot_weather_cat}
\title{Plot Weather Information on Map}
\usage{
plot_weather_cat(id_vector, type, label = TRUE)
}
\arguments{
\item{id_vector}{a vector of multiple airport codes}

\item{type}{a character of one categorical weather element}

\item{label}{lable= TRUE: label airport code on the plot}
}
\description{
plot a categorical element of an weather informaton of multiple airports on map
}
\examples{
plot_weather_cat(c("KAMW", "KAIO", "KCID", "KCNC"),
           type = "weather",
           label = TRUE)
}
