% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather_num.R
\name{plot_weather_num}
\alias{plot_weather_num}
\title{Plot Weather Information on Map}
\usage{
plot_weather_num(id_vector, type, label = TRUE)
}
\arguments{
\item{id_vector}{a vector of multiple airport codes}

\item{type}{a character of one numerical weather element}

\item{label}{lable= TRUE: label airport code on the plot}
}
\description{
plot a numeric element of an weather informaton of multiple airports on map
}
\examples{
plot_weather_num(c("KAMW", "KAIO", "KCID", "KCNC"),
           type = "temp_f",
           label = TRUE)
}
