% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather_us.R
\name{plot_weather_us}
\alias{plot_weather_us}
\title{Plot an Numeric Element of All US Airport in Contour Lines}
\usage{
plot_weather_us(type, you.long, you.lat, bin)
}
\arguments{
\item{type}{a vector of weather elements}

\item{you.long}{The longitude of your location}

\item{you.lat}{The latitude of your location}

\item{bin}{the binwidth of the contour lines}
}
\description{
This function plot an numeric element of all US airport in contour lines
}
\examples{
plot_weather_us(type = "temp_c",
you.long = -100, you.lat = 35,bin = 0.5)
}
