% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{plot_weather}
\alias{plot_weather}
\title{Plot Weather Information on Map}
\usage{
plot_weather(id_vector, type, label = T, number = T)
}
\arguments{
\item{id_vector}{a vector of multiple airport codes}

\item{type}{a vector of weather elements}
}
\description{
plot an element of an weather informaton of multiple airports on map
}
\details{
label = TRUE: label airport code on the plot

number = TRUE: this element is numeric
}
\examples{
plot_weather(c("KAMW", "KAIO", "KCID", "KCNC"),
           type = "weather",
           label = T, number = F)
}
